% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{april_1_snotel_data}
\alias{april_1_snotel_data}
\title{april_1_snotel_data}
\format{
An object of class \code{data.frame} with 3768 rows and 23 columns.
}
\usage{
april_1_snotel_data
}
\description{
All Snotel station data for april 1st in utah with variables of Lat, Lon,
  Elev, and PRISM variables
}
\keyword{datasets}
