% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend_raster.R
\name{blend_raster}
\alias{blend_raster}
\title{Blending Satellite and land-based rasters with weights}
\usage{
blend_raster(raster_sate, raster_land, weights)
}
\arguments{
\item{raster_sate}{A raster object of the satellite source estimates}

\item{raster_land}{A raster object of estimates from the the land-based
observation}

\item{weights}{A raster object with weights}
}
\value{
a data frame with the columns with the variables needed to
  make predictions of the model.
}
\description{
Blending Satellite and land-based rasters with weights
}
\details{
This function combines gridded product by taking in two rasters and
  the weights applied to both rasters.
}
