% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_raster.R
\name{df_to_raster}
\alias{df_to_raster}
\title{Generate a raster from data frame}
\usage{
df_to_raster(model, data_frame, raster_template)
}
\arguments{
\item{model}{the model object that will be used to predict.}

\item{data_frame}{a data frame with the columns of the model that will
be used to create the raster}

\item{raster_template}{a stars object that is contains the area and
projection wanted.}
}
\value{
a stars object that contains the predictions of the model
}
\description{
Generate a raster from data frame
}
\details{
This function creates gridded product by taking in a model_data,
  raster template, a model object or model x and y's, path to prism variables,

Takes a model and a data frame of climate variables that will be
  used to predict estimates in a raster or grid format. This will return
  a stars object of the estimates.
}
