% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_actionid_stations.R
\name{download_actionid_stations}
\alias{download_actionid_stations}
\title{Download the Stations with ActionID}
\usage{
download_actionid_stations()
}
\value{
a dataframe with station names with action Ids
}
\description{
Download the Stations with ActionID
}
\details{
This function helps with getting stations that have an Action Id.
  The stations with an Action Id are in both the NRCS and GHCND.
}
