% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all_ghcnd_stations.R
\name{download_all_ghcnd_stations}
\alias{download_all_ghcnd_stations}
\title{Download all .dly weather data from the web.}
\usage{
download_all_ghcnd_stations(directory = ".", keep_compressed = FALSE)
}
\arguments{
\item{directory}{The directory to download the .dly files. Default is the
current working directory.}

\item{keep_compressed}{If TRUE, the compressed file is retained in the
specified directory. The default false removes the tar file after
unzipping.}
}
\value{
Uncompressed ghcnd_all folder and ghcnd-version.txt in the supplied
directory.
}
\description{
This function is designed to facilitate faster creation of data sets in
the functions get_station_data and get_state_data functions.
This is done by downloading all .dly files available, then
using the downloaded folder as the "source" instead of the web source.
}
