% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_snodas.R
\name{download_snodas}
\alias{download_snodas}
\title{Download SNODAS Data maps}
\usage{
download_snodas(
  dates = c("2021-04-01", "2022-4-1"),
  masked = TRUE,
  overwrite = TRUE,
  remove_zip = TRUE,
  data_saved = c("swe", "SP", "SD", "SPT", "bss", "melt", "SPS", "NSP"),
  out_dir = paste0(getwd(), "/snodas_data"),
  GTiff = TRUE
)
}
\arguments{
\item{dates}{date with YYYY-MM-dd written as a six digit character string}

\item{masked}{if TRUE, the maps will be from the masked
if FALSE, the maps returned will be unmasked.}

\item{overwrite}{if TRUE, if the file already exists it will be overwritten}

\item{remove_zip}{if TRUE, the temporary folder with all the zipped folder
will be deleted.}

\item{data_saved}{data_saved is the type of maps wanted. There are 8 options:
"SWE" for Snow-water-equivalent. "
"SP" for Snow precipatation 24 hrs
"SD" for Modeled Snow Depth
"SPT" for Snowpack average temperature
"BSS" for Modeled blowing snow sublimation rate
"MELT" for Modeled melt rate
"SPS" for Modeled snowpack sublimation rate
"NSP" for Non-snow precipitation}

\item{out_dir}{out_dir is the name of a output directory or folder that will
be created in the working directory. The User can input a name for the
in which the data will be saved in.}

\item{GTiff}{a logical variable that will save the maps of interest as a
gtif object.}
}
\value{
a list of star objects with elements corresponding to the
  input argument data_saved.
}
\description{
Download SNODAS Data maps
}
\details{
This function creates a permanent folder and downloads data
  from the SNODAS website for the selected date(s), map type(s), and data map
  that the user provides. This data is stored in the permanent folder.
  The function unzips the selected data files and stores the specified
  maps into a list of star objects.
}
