% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_snodas_urls.R
\name{download_snodas_urls}
\alias{download_snodas_urls}
\title{Download Snodas Urls to a specific folder}
\usage{
download_snodas_urls(dates, masked, urls)
}
\arguments{
\item{dates}{date with YYYY-MM-dd written as a six digit character string}

\item{masked}{if TRUE, the maps will be from the masked
if FALSE, the maps returned will be unmasked.}

\item{urls}{a vector of urls that are from snodas and used to}
}
\value{
a vector of urls that are in the same order as the dates given.
}
\description{
Download Snodas Urls to a specific folder
}
\details{
This function takes dates in the form of YYYY-MM-DD and returns a
  vector of urls that can be used to download the information for the
  specific dates given.
}
