% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_to_df.R
\name{gam_to_df}
\alias{gam_to_df}
\title{Generate a data frame from a generalized additive model}
\usage{
gam_to_df(
  model_data,
  raster_template,
  model_x,
  model_y,
  coords = c("LONGITUDE", "LATITUDE"),
  path_to_prism
)
}
\arguments{
\item{model_data}{the data in which the model is create. The variables of
'slope', 'aspect', and 'elevation' need to be lowercase.}

\item{raster_template}{a data frame with the columns of the model that will
be used to create the model.}

\item{model_x}{a character vector where each entry symbolizes a column
in the model_data data frame.  These will be the variables used to try
to explain the Independent variable.}

\item{model_y}{a character vector of the Independent variable.}

\item{coords}{the LON/LAT column names of the data.  This will be used to
in the model no matter what.}

\item{path_to_prism}{a character vector that has the file path
to the PRISM files.}
}
\value{
a data frame with the columns with the variables needed to
  make predictions of the model.
}
\description{
Generate a data frame from a generalized additive model
}
\details{
This function creates gridded product by taking in a model_data,
  raster template, a model object or model x and y's, path to prism variables,

This function creates a permanent folder and downloads data
  from the SNODAS website for the selected date(s), map type(s), and data map
  that the user provides. This data is stored in the permanent folder.
  The function unzips the selected data files and stores the specified
  maps into a list of star objects.
}
