% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_to_raster.R
\name{gam_to_raster}
\alias{gam_to_raster}
\title{Generate a raster from a generalized additive model}
\usage{
gam_to_raster(model_data, raster_template, model = NA, path_to_prism)
}
\arguments{
\item{model_data}{the data in which the model is created}

\item{raster_template}{a data frame with the columns of the model that will
be used to create the model.}

\item{model}{to apply to the data and predict.}

\item{path_to_prism}{a character vector that has the path to the PRISM files}
}
\value{
a stars object that contains the predictions of the model
}
\description{
Generate a raster from a generalized additive model
}
\details{
This function creates a permanent folder and downloads data
from the SNODAS website for the inputed date(s), map type(s), and data map
that the user provides. This data is stored in the permanent folder.
The function unzips the selected data files and stores the specified
maps into a list and combines them all into a RasterBrick.

This function creates a permanent folder and downloads data
  from the SNODAS website for the selected date(s), map type(s), and data map
  that the user provides. This data is stored in the permanent folder.
  The function unzips the selected data files and stores the specified
  maps into a list of star objects.
}
