% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_to_raster.R
\name{get_prism_info}
\alias{get_prism_info}
\title{Helper function in GAM to Raster}
\usage{
get_prism_info(x)
}
\arguments{
\item{x}{a character vector with model parameters}
}
\value{
a character vector with model parameters
}
\description{
Helper function in GAM to Raster
}
\details{
helps for naming and reading in files for raster or star objects
}
