% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snodas_daily.R
\name{get_snodas_daily}
\alias{get_snodas_daily}
\title{Download Daily SNODAS Data maps}
\usage{
get_snodas_daily(
  dates = c("2010-01-01", "2012-4-15"),
  masked = TRUE,
  data = c("swe", "SP", "SD", "SPT", "bss", "melt", "SPS", "NSP"),
  permanent = "permanent folder"
)
}
\arguments{
\item{dates}{date with YYYY-MM-dd written as a six digit character string}

\item{masked}{if TRUE, the maps will be from the masked
if FALSE, the maps returned will be unmasked.}

\item{data}{data is the type of maps wanted. There are 8 options:
"SWE" for Snow-water-equivalent. "
"SP" for Snow precipatation 24 hrs
"SD" for Modeled Snow Depth
"SPT" for Snowpack average temperature
"BSS" for Modeled blowing snow sublimation rate
"MELT" for Modeled melt rate
"SPS" for Modeled snowpack sublimation rate
"NSP" for Non-snow precipitation}

\item{permanent}{permanent is the name of a permanent folder that will
be created in the working directory. The User can input a different name
for the permanent folder.}
}
\value{
a brick of rasters with elements corresponding to the
  input vaiable "data".
}
\description{
Download Daily SNODAS Data maps
}
\details{
This function creates a permanent folder and downloads data
from the SNODAS website for the inputed date(s), map type(s), and data map
that the user provides. This data is stored in the permanent folder.
The function unzips the selected data files and stores the specified
maps into a list and combines them all into a RasterBrick.
}
