% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snodas_urls.R
\name{get_snodas_urls}
\alias{get_snodas_urls}
\title{Get URLS for each specific date}
\usage{
get_snodas_urls(dates, masked)
}
\arguments{
\item{dates}{date with YYYY-MM-dd written as a six digit character string}

\item{masked}{if TRUE, the maps will be from the masked
if FALSE, the maps returned will be unmasked.}
}
\value{
a vector of urls that are in the same order as the dates given.
}
\description{
Get URLS for each specific date
}
\details{
This function takes dates in the form of YYYY-MM-DD and returns a
  vector of urls that can be used to download the information for the
  specific dates given.
}
