% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ua_yearly.R
\name{get_ua_yearly}
\alias{get_ua_yearly}
\title{Download yearly University of Arizona 4km maps}
\usage{
get_ua_yearly(year = 2004, permanent = "permanent folder")
}
\arguments{
\item{year}{with YYYY, written as a 4 digit character string}

\item{permanent}{permanent is the name of a permanent folder that will
be created in the working directory. The User can input a different name
for the permanent folder.}
}
\value{
a brick of rasters with elements corresponding to the
  input vaiable "SWE" or Snowdepth.
}
\description{
This function creates a permanent folder and downloads data
  from the SNODAS website for the inputed date(s), map type(s), and data map
  that the user provides. This data is stored in the permanent folder.
  The function unzips the selected data files and stores the specified
  maps into a list and combines them all into a RasterBrick.
}
