% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghcnd_stations}
\alias{ghcnd_stations}
\title{ghcnd_stations}
\format{
An object of class \code{data.frame} with 69139 rows and 13 columns.
}
\usage{
ghcnd_stations
}
\description{
A dataset of all the ghcnd_stations
}
\keyword{datasets}
