% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_to_density_stars.R
\name{points_to_density_stars}
\alias{points_to_density_stars}
\title{Generate a raster from the density of spatial points}
\usage{
points_to_density_stars(
  sp_points = april_df_ghcnd[[17]],
  coords = c("LONGITUDE", "LATITUDE"),
  raster_template = snodas_maps_utah[[17]],
  sigma = 15000,
  max_weight = 0.5,
  flat_crs = "+proj=utm + zone=12 + datum=WGS84"
)
}
\arguments{
\item{sp_points}{a data frame of spatial points that must include the columns
longitude, latitude, and other variables}

\item{coords}{the column names of the coordinates longitude and latitude.}

\item{raster_template}{a template that the density raster should follow}

\item{sigma}{changes the distance in which to consider}

\item{max_weight}{a numeric value between between (0, 1].}

\item{flat_crs}{a flat (2D) projection that the user specifies.  The user
should consider the location of interest to determine the best projection.}
}
\value{
a stars object that has the density of the spatial point data.
}
\description{
Generate a raster from the density of spatial points
}
\details{
This function creates a permanent folder and downloads data
from the SNODAS website for the inputed date(s), map type(s), and data map
that the user provides. This data is stored in the permanent folder.
The function unzips the selected data files and stores the specified
maps into a list and combines them all into a RasterBrick.
}
