% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_to_raster.R
\name{rm_asp_slop_elev}
\alias{rm_asp_slop_elev}
\title{Helper function in GAM to Raster}
\usage{
rm_asp_slop_elev(x)
}
\arguments{
\item{x}{a character vector with model parameters}
}
\value{
a character vector with model parameters
}
\description{
Helper function in GAM to Raster
}
\details{
helps get rid of aspect slope and elevation in one stop because
  they all use the same PRISM file.
}
