% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_raster.R
\name{df_to_raster}
\alias{df_to_raster}
\title{Generate a raster from data frame}
\usage{
df_to_raster(model, df)
}
\arguments{
\item{model}{the model object that will be used to predict.}

\item{df}{a data frame with the columns of the model that will
be used to create the raster}
}
\value{
a stars object that contains the predictions of the model
}
\description{
Generate a raster from data frame
}
\details{
This function creates gridded product by taking in a model_data,
  raster template, a model object or model x and y's, path to prism variables,

This function creates a permanent folder and downloads data
  from the SNODAS website for the selected date(s), map type(s), and data map
  that the user provides. This data is stored in the permanent folder.
  The function unzips the selected data files and stores the specified
  maps into a list of star objects.
}
